Unify RubyLexer.isSpace() with StringSupport.isAsciiSpace()
* Allow int argument, it's the same semantics for 7-bit integers anyway.
* Use character constants for readability.

+++ src/main/java/org/truffleruby/core/string/StringNodes.java
@@ -1357,7 +1357,7 @@ public abstract class StringNodes {
             // Check the first code point to see if it's a space. In the case of strings without leading spaces,
             // this check can avoid having to materialize the entire byte[] (a potentially expensive operation
             // for ropes) and can avoid having to compile the while loop.
-            if (noopProfile.profile(!StringSupport.isAsciiSpace((byte) firstCodePoint))) {
+            if (noopProfile.profile(!StringSupport.isAsciiSpace(firstCodePoint))) {
                 return nil();
             }
 
@@ -1470,7 +1470,7 @@ public abstract class StringNodes {
             // Check the last code point to see if it's a space or NULL. In the case of strings without leading spaces,
             // this check can avoid having to materialize the entire byte[] (a potentially expensive operation
             // for ropes) and can avoid having to compile the while loop.
-            final boolean willStrip = lastCodePoint == 0x00 || StringSupport.isAsciiSpace((byte) lastCodePoint);
+            final boolean willStrip = lastCodePoint == 0x00 || StringSupport.isAsciiSpace(lastCodePoint);
             if (noopProfile.profile(!willStrip)) {
                 return nil();
             }
